Organises tests into folders
