+++ .phpunit.result.cache
@@ -1,0 +1,1 @@
+C:37:"PHPUnit\Runner\DefaultTestResultCache":4825:{a:2:{s:7:"defects";a:1:{s:80:"RicorocksDigitalAgency\Soap\Tests\GlobalHooksTest::hooks_run_even_without_faking";i:3;}s:5:"times";a:44:{s:80:"RicorocksDigitalAgency\Soap\Tests\IntelligentBuilderTest::it_can_handle_an_array";d:0.002;s:86:"RicorocksDigitalAgency\Soap\Tests\IntelligentBuilderTest::it_can_handle_a_nested_array";d:0.002;s:78:"RicorocksDigitalAgency\Soap\Tests\IntelligentBuilderTest::it_can_handle_a_node";d:0.003;s:99:"RicorocksDigitalAgency\Soap\Tests\IntelligentBuilderTest::it_can_handle_a_node_with_only_attributes";d:0.003;s:103:"RicorocksDigitalAgency\Soap\Tests\IntelligentBuilderTest::it_can_handle_a_node_with_attributes_and_body";d:0.002;s:84:"RicorocksDigitalAgency\Soap\Tests\IntelligentBuilderTest::it_can_handle_nested_nodes";d:0.002;s:69:"RicorocksDigitalAgency\Soap\Tests\SoapClassTest::it_can_obtain_a_wsdl";d:0.003;s:76:"RicorocksDigitalAgency\Soap\Tests\SoapClassTest::it_can_call_a_SOAP_function";d:0.002;s:65:"RicorocksDigitalAgency\Soap\Tests\SoapClassTest::it_can_use_nodes";d:0.002;s:76:"RicorocksDigitalAgency\Soap\Tests\SoapClassTest::it_can_forward_method_calls";d:0.002;s:124:"RicorocksDigitalAgency\Soap\Tests\SoapClassTest::it_can_include_an_array_at_the_root_when_specified_using_the_include_method";d:0.007;s:122:"RicorocksDigitalAgency\Soap\Tests\SoapClassTest::it_can_include_a_node_at_the_root_when_specified_using_the_include_method";d:0.003;s:92:"RicorocksDigitalAgency\Soap\Tests\SoapClassTest::it_only_includes_if_the_method_name_matches";d:0.002;s:70:"RicorocksDigitalAgency\Soap\Tests\SoapFakeTest::it_can_record_requests";d:0.003;s:68:"RicorocksDigitalAgency\Soap\Tests\SoapFakeTest::it_can_fake_requests";d:0.003;s:78:"RicorocksDigitalAgency\Soap\Tests\SoapFakeTest::it_can_fake_specific_endpoints";d:0.005;s:71:"RicorocksDigitalAgency\Soap\Tests\SoapFakeTest::it_can_handle_wildcards";d:0.005;s:80:"RicorocksDigitalAgency\Soap\Tests\SoapFakeTest::it_can_handle_multiple_wildcards";d:0.003;s:69:"RicorocksDigitalAgency\Soap\Tests\SoapFakeTest::it_can_handle_methods";d:0.003;s:96:"RicorocksDigitalAgency\Soap\Tests\GlobalHooksTest::it_can_perform_global_before_requesting_hooks";d:0.003;s:95:"RicorocksDigitalAgency\Soap\Tests\GlobalHooksTest::it_can_perform_global_after_requesting_hooks";d:0.002;s:80:"RicorocksDigitalAgency\Soap\Tests\GlobalHooksTest::hooks_run_even_without_faking";d:0.003;s:106:"RicorocksDigitalAgency\Soap\Tests\SoapFakeTest::it_can_handle_multiple_methods_declared_by_a_pipe_operator";d:0.003;s:75:"RicorocksDigitalAgency\Soap\Tests\Fake\SoapFakeTest::it_can_record_requests";d:0.003;s:73:"RicorocksDigitalAgency\Soap\Tests\Fake\SoapFakeTest::it_can_fake_requests";d:0.002;s:83:"RicorocksDigitalAgency\Soap\Tests\Fake\SoapFakeTest::it_can_fake_specific_endpoints";d:0.002;s:76:"RicorocksDigitalAgency\Soap\Tests\Fake\SoapFakeTest::it_can_handle_wildcards";d:0.003;s:85:"RicorocksDigitalAgency\Soap\Tests\Fake\SoapFakeTest::it_can_handle_multiple_wildcards";d:0.003;s:74:"RicorocksDigitalAgency\Soap\Tests\Fake\SoapFakeTest::it_can_handle_methods";d:0.005;s:111:"RicorocksDigitalAgency\Soap\Tests\Fake\SoapFakeTest::it_can_handle_multiple_methods_declared_by_a_pipe_operator";d:0.004;s:76:"RicorocksDigitalAgency\Soap\Tests\Fake\MethodFakeTest::it_can_handle_methods";d:0.005;s:113:"RicorocksDigitalAgency\Soap\Tests\Fake\MethodFakeTest::it_can_handle_multiple_methods_declared_by_a_pipe_operator";d:0.003;s:102:"RicorocksDigitalAgency\Soap\Tests\Hooks\GlobalHooksTest::it_can_perform_global_before_requesting_hooks";d:0.003;s:101:"RicorocksDigitalAgency\Soap\Tests\Hooks\GlobalHooksTest::it_can_perform_global_after_requesting_hooks";d:0.003;s:86:"RicorocksDigitalAgency\Soap\Tests\Hooks\GlobalHooksTest::hooks_run_even_without_faking";d:0.002;s:133:"RicorocksDigitalAgency\Soap\Tests\Inclusions\IncludeTest::it_can_include_an_array_at_the_root_when_specified_using_the_include_method";d:0.006;s:131:"RicorocksDigitalAgency\Soap\Tests\Inclusions\IncludeTest::it_can_include_a_node_at_the_root_when_specified_using_the_include_method";d:0.003;s:101:"RicorocksDigitalAgency\Soap\Tests\Inclusions\IncludeTest::it_only_includes_if_the_method_name_matches";d:0.003;s:89:"RicorocksDigitalAgency\Soap\Tests\Building\IntelligentBuilderTest::it_can_handle_an_array";d:0.038;s:95:"RicorocksDigitalAgency\Soap\Tests\Building\IntelligentBuilderTest::it_can_handle_a_nested_array";d:0.003;s:87:"RicorocksDigitalAgency\Soap\Tests\Building\IntelligentBuilderTest::it_can_handle_a_node";d:0.003;s:108:"RicorocksDigitalAgency\Soap\Tests\Building\IntelligentBuilderTest::it_can_handle_a_node_with_only_attributes";d:0.002;s:112:"RicorocksDigitalAgency\Soap\Tests\Building\IntelligentBuilderTest::it_can_handle_a_node_with_attributes_and_body";d:0.002;s:93:"RicorocksDigitalAgency\Soap\Tests\Building\IntelligentBuilderTest::it_can_handle_nested_nodes";d:0.002;}}}
