+++ src/Facades/Soap.php
@@ -19,10 +19,10 @@ use RicorocksDigitalAgency\Soap\Request\Request;
  * @method static \RicorocksDigitalAgency\Soap\Soap beforeRequesting(callable $hook)
  * @method static \RicorocksDigitalAgency\Soap\Soap afterRequesting(callable $hook)
  * @method static void fake(array|Closure $callable = null)
- * @method static assertSentCount($count)
- * @method static assertNothingSent()
- * @method static assertSent(callable $callback)
- * @method static assertNotSent(callable $callback)
+ * @method static void assertSentCount($count)
+ * @method static void assertNothingSent()
+ * @method static void assertSent(callable $callback)
+ * @method static void assertNotSent(callable $callback)
  */
 class Soap extends Facade
 {
